remove useSearchQueryAttributes hook
