+++ src/@next/components/organisms/ProductVariantPicker/types.ts
@@ -11,4 +11,5 @@ export interface IProps {
   selectSidebar?: boolean;
   selectSidebarTarget?: HTMLElement | null;
   queryAttributes: Record<string, string>;
+  updateUrlWithAttributes: (slug: string, value: string) => void;
 }
