+++ src/@next/hooks/index.ts
@@ -7,4 +7,3 @@ export * from "./useProductVariantsAttributesValuesSelection";
 export * from "./useSelectableProductVariantsAttributeValues";
 export * from "./useCheckoutStepState";
 export * from "./useCheckoutStepFromPath";
-export * from "./useSearchQueryAttributes";
