+++ src/views/Product/Page.tsx
@@ -26,12 +26,10 @@ const populateBreadcrumbs = product => [
   },
 ];
 
-const Page: React.FC<IProps & { queryAttributes: Record<string, string> }> = ({
-  add,
-  product,
-  items,
-  queryAttributes,
-}) => {
+const Page: React.FC<IProps & {
+  queryAttributes: Record<string, string>;
+  updateUrlWithAttributes: (slug: string, value: string) => void;
+}> = ({ add, product, items, queryAttributes, updateUrlWithAttributes }) => {
   const productGallery: React.RefObject<HTMLDivElement> = React.useRef();
 
   const [variantId, setVariantId] = React.useState("");
@@ -61,6 +59,7 @@ const Page: React.FC<IProps & { queryAttributes: Record<string, string> }> = ({
       queryAttributes={queryAttributes}
       addToCart={add}
       setVariantId={setVariantId}
+      updateUrlWithAttributes={updateUrlWithAttributes}
     />
   );
